server/requirements.txt
alembic                    1.11.1
antiorm                    1.2.1
blinker                    1.6.2
certifi                    2023.5.7
charset-normalizer         3.1.0
click                      8.1.3
colorama                   0.4.6
db                         0.1.1
db-sqlite3                 0.0.1
exceptiongroup             1.1.1
Flask                      2.3.2
Flask-Cors                 3.0.10
Flask-JWT-Extended         4.4.4
Flask-Login                0.6.2
Flask-Mail                 0.9.1
flask-mailman              0.3.0
Flask-SQLAlchemy           3.0.3
greenlet                   2.0.2
idna                       3.4
importlib-metadata         6.6.0
iniconfig                  2.0.0
itsdangerous               2.1.2
Jinja2                     3.1.2
Mako                       1.2.4
MarkupSafe                 2.1.2
mkdocs-material-extensions 1.1.1
packaging                  23.1
pip                        21.2.4
pluggy                     1.0.0
psycopg2                   2.9.6
PyJWT                      2.7.0
pytest                     7.3.1
python-dotenv              1.0.0
requests                   2.30.0
setuptools                 58.1.0
six                        1.16.0
SQLAlchemy                 2.0.13
tomli                      2.0.1
typing_extensions          4.5.0
urllib3                    2.0.2
Werkzeug                   2.3.4
zipp                       3.15.0